Add files via upload
These two scripts should replace charucoCalibration.py when complete. They complete the same fundamental task, but they separate the collection of the pertinent information (corners, ids, frame size, charuco board object) from the processing of this information. This is done by storing this information using pickle. By moving this file to a separate computer, it should be possible for a different computer to perform the computationally costly calibration. This would be very useful when calibrating Raspberry Pi cameras. The cameras can be used by the Pi to collect data, but the actual calibration could be completed by a more powerful computer.